moved to using API Controller
