+++ app/Http/Controllers/API/FlowController.php
@@ -1,0 +1,88 @@
+<?php
+
+namespace App\Http\Controllers\API;
+
+use Illuminate\Http\Request;
+use App\Http\Controllers\Controller;
+
+class FlowController extends Controller
+{
+    /**
+     * Display a listing of the resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function index()
+    {
+        // todo: move to query object
+        // todo: use the request to get the User in the session
+
+        // move to a service that takes the command
+        $results = Flow::all();
+        // returns dataset
+
+        // todo: use fractal / transformer
+
+        return response()->json([
+            'message' => 'Flow index route',
+            'data' => $results,
+        ]);
+    }
+
+    /**
+     * Store a newly created resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\Response
+     */
+    public function store(Request $request)
+    {
+        $title = $request->input('title');
+
+        return response()->json([
+            'data' => "hit store route with new title $title",
+        ]);
+    }
+
+    /**
+     * Display the specified resource.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function show(int $id)
+    {
+        dump('>>> flow id', $id);
+//        Flow::where('id',)[
+        return response()->json([
+            'message' => 'Flow show route',
+        ]);
+    }
+
+    /**
+     * Update the specified resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function update(Request $request, $id)
+    {
+        return response()->json([
+            'message' => 'Flow update route',
+        ]);
+    }
+
+    /**
+     * Remove the specified resource from storage.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function destroy($id)
+    {
+        return response()->json([
+            'message' => 'Flow destroy route',
+        ]);
+    }
+}
